skills/doc-to-markdown/pyproject.toml
[project]
name = "doc-to-markdown"
version = "0.1.0"
description = "Convert documents to markdown using marker"
requires-python = ">=3.11"
dependencies = [
    "marker-pdf>=1.10.0",
    "weasyprint>=68.0",
    "openpyxl>=3.1.5",
]

[project.scripts]
convert = "convert:main"
